requirements.txt
paramiko~=2.11.0
adal==1.2.0
aioredis==1.2.0
amqp==2.3.2
APScheduler==3.6.3
asgiref==2.3.2
asn1crypto==0.23.0
async-timeout==3.0.1
attrs==21.4.0
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
backports.csv==1.0.7
bcrypt==3.1.7
beautifulsoup4==4.6.0
billiard==3.5.0.4
cachetools==3.0.0
celery==4.2.1
celery-once==2.0.0
certifi==2017.7.27.1
cffi==1.13.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
#client==0.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==2.5
daphne==2.2.5
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-apscheduler==0.3.0
django-auth-ldap==1.7.0
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-extensions==1.7.9
django-filter==2.2.0
#django-filters==0.2.1
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.2.0
django-jsonfield==1.0.1
django-mysql==3.0.0.post1
django-redis==4.11.0
django-reversion==3.0.4
django-simple-history==2.7.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
dwebsocket==0.5.10
et-xmlfile==1.0.1
ez-setup==0.9
fabric2==2.4.0
flower==0.9.2
future==0.17.1
google-auth==1.6.1
greenlet==0.4.15
gunicorn==19.9.0
hiredis==1.0.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
importlib-metadata==0.23
incremental==17.5.0
invoke==1.2.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10
kombu==4.2.1
kubernetes==8.0.0
ldap==1.0.2
ldap3==2.4.1
Markdown==3.0.1
MarkupSafe==1.1.0
more-itertools==7.2.0
msgpack==0.6.1
msgpack-python==0.5.6
multi-key-dict==2.0.3
mysqlclient==1.3.13
nose==1.3.7
numpy==1.17.3
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.25.3
pbr==5.2.1
permission==0.4.1
pexpect==4.3.1
Pillow==5.4.1
ptyprocess==0.5.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyHamcrest==1.9.0
PyJWT==1.6.4
pyldap==3.0.0.post1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.7.5
python-jenkins==1.7.0
python-kubernetes==0.2
python-ldap==3.1.0
python-magic==0.4.15
python-nmap==0.6.1
pytz==2017.2
PyYAML==3.13
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel.yaml==0.15.80
simplejson==3.13.2
six==1.11.0
South==1.0.2
SQLAlchemy==1.3.5
sqlparse==0.3.0
tablib==0.13.0
threadpool==1.3.2
tornado==5.0.1
Twisted==18.9.0
txaio==18.8.1
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
websocket-client==0.54.0
xlrd==1.2.0
XlsxWriter==1.1.6
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.6.0

ufile~=3.2.7
jenkinsapi~=0.3.13